Finish moving app
